requirements.txt
numpy
pandas
opencv-python
pyrealsense2
pyserial
geopy
pillow